requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
bidict==0.22.0
boto3==1.24.17
botocore==1.27.17
CacheControl==0.12.11
cachetools==4.2.4
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==3.4.8
cycler==0.11.0
dask==2022.2.0
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.4.0
flatbuffers==2.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
httptools==0.4.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==4.4.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
locket==1.0.0
Markdown==3.3.7
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.4
multidict==5.2.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
partd==1.2.0
Pillow==9.1.1
prompt-toolkit==2.0.10
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyTelegramBotAPI==4.6.0
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.2.1
rasa-sdk==3.2.0
redis==3.5.3
regex==2022.4.24
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rocketchat-API==1.25.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sanic==21.12.1
Sanic-Cors==2.0.1
sanic-jwt==1.7.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentry-sdk==1.3.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.39
tabulate==0.8.10
tarsafe==0.0.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.3
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-text==2.7.3
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
toolz==0.11.2
tqdm==4.64.0
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==5.3.0
urllib3==1.26.9
uvloop==0.16.0
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
